Fix reading of IXF files data records.
From the specs about the IXFDRID field:

  The record identifier,  which identifies a particular  D record within
  the sequence of  several D records contributing to a  row of data. For
  the first D  record in a row of  data, this field has a  value of one;
  for the second  D record in a row  of data, this field has  a value of
  two, and  so on.  In each row  of data, all  the D  record identifiers
  called out in the C records must actually exist.

In this commit we make it so that the IXF reader is able to switch to
the next data record when needed. To get there, the internal APIs needed
to change, and that triggered in turn a review of the main API, which
has been improved.

Read: the main API just got broken, you need to change your code if you
use this lib.