Implement reading clob content.
For parsing the LOB Location Specifier, the blog post at the following
URL did help a lot, the IBM docs not mentionning what it should be
expected to look like...

  https://www.toadworld.com/platforms/ibmdb2/w/wiki/7545.exporting-lobs-and-identity-columns

The binary LOBs support is not expected to work fully yet for lack of
sample data. If you need this support, you should know how to help
getting it: send sample data my way!